Refactoring for Usability (#4)
* Refactoring for Usability

settingsPath is now facilityConfig to make way for a future clientConfig XML to universalize the facility config files.

CLI arguments now follow standard CLI convention (as defined by http://docopt.org/)

Added a default to ver. 1.0.0 if no version is found... This is mostly a hack to make things run correctly locally, but could apply to future version checking

* Allow /S Screensaver Call

Because windows is weird and bad

* Abandon sanity, return to Windows

/C, /P, /S screensaver commands work again now. Also, you can open files by dragging them onto the EXE (or, via shortcut, simply passing the filepath).

* Reorganize start sequence

Cleaned up the startup sequence to allow for more versatility with shortcut/CLI arguments.  Screensavers can now be passed a particular config file to run, and users who fail to select a facility config file on startup will be prompted to create a new one (or not).

* tweak arguments

* actually allow the user to create a new config

* restore writing the version number to the output

* Slightly clearer fake version string

Co-authored-by: dg <dennis.graiani@gmail.com>